+++ capi/base/c_api.cpp
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 24/02/2023
+// Last Modified: 03/03/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -479,10 +479,10 @@ void AUTDSetMode(void* const handle, const uint8_t mode) {
       *wrapper << autd3::legacy_mode;
       break;
     case 1:
-      *wrapper << autd3::normal_mode;
+      *wrapper << autd3::advanced_mode;
       break;
     case 2:
-      *wrapper << autd3::normal_phase_mode;
+      *wrapper << autd3::advanced_phase_mode;
       break;
     default:
       break;