+++ include/autd3/driver/operation/operation.hpp
@@ -3,7 +3,7 @@
 // Created Date: 06/01/2023
 // Author: Shun Suzuki
 // -----
-// Last Modified: 16/01/2023
+// Last Modified: 03/03/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -17,8 +17,8 @@ namespace autd3::driver {
 
 struct Mode {};
 struct Legacy : Mode {};
-struct Normal : Mode {};
-struct NormalPhase : Mode {};
+struct Advanced : Mode {};
+struct AdvancedPhase : Mode {};
 
 template <typename T>
 inline constexpr bool is_mode_v = std::is_base_of_v<Mode, T>;