+++ julia/example/simulator.jl
@@ -3,7 +3,7 @@
 # Created Date: 30/12/2020
 # Author: Shun Suzuki
 # -----
-# Last Modified: 02/02/2023
+# Last Modified: 03/03/2023
 # Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 # -----
 # Copyright (c) 2020 Hapis Lab. All rights reserved.
@@ -26,7 +26,7 @@ function main()
 
         cnt = Controller(geometry, link)
 
-        cnt.to_normal()
+        cnt.to_advanced()
         for tr in cnt.geometry()
             tr.set_frequency(70e3)
         end